+++ gpu_manager.cpp
@@ -22,9 +22,9 @@ gpu_manager::gpu_manager(int num_batches, long batch_size, int dimension){
     for(int i=0; i<this->num_devices; i++)
         this->list_of_users.push_back(new single_gpu_manager(i, device_num_batches, batch_size, dimension));
 
-    result_vector = (float*)malloc(num_batches * batch_size * sizeof(float)); // new double[data_size];
+    result_vector = (float*)malloc(num_batches * batch_size * sizeof(float)); // new float[data_size];
     if(result_vector == nullptr)
-        mem_error("result_vector", "vectmatmul", device_num_batches * batch_size, "double");
+        mem_error("result_vector", "vectmatmul", device_num_batches * batch_size, "float");
 
 }
 
@@ -76,7 +76,7 @@ void gpu_manager::search() {
 
 void gpu_manager::top_k(int k){
       auto start = std::chrono::steady_clock::now();
-      std::vector<double> myvector (result_vector, result_vector + device_num_batches * batch_size);
+      std::vector<float> myvector (result_vector, result_vector + device_num_batches * batch_size);
       std::partial_sort (myvector.begin(), myvector.begin()+k, myvector.end(), [](int a, int b) {return a > b; });
       auto stop = std::chrono::steady_clock::now();
       double time_sec = std::chrono::duration_cast<std::chrono::nanoseconds>(stop - start).count() * 1e-9;
